Add OWNERS to libhardware_legacy.
libhardware_legacy is the legacy deprecated version
of the legacy and deprecated library libhardware.
In it are defined vendor/system interfaces. One common
pattern is for the following call flow to happen:

<vendor code>
  -> legacy interface
      -> HIDL
          -> legacy interface
              -> <framework client code>

Because of this, it's possible to circumvent the stable ABI
provided in HIDL by accidentally passing through data which
is not represented in the vendor interface. This causes
coupling of system/vendor components by way of an undeclared
ABI.

This initial OWNERS file was created with the following groups:
    <HIDL specific players>
    <common committers> intersect <familiarity with this problem>

Bug: 70526789
Bug: 37280010
Test: N/A

Merged-In: Icebeb958b41cb505d76341f3d53a3e175a3c92bd
Change-Id: Icebeb958b41cb505d76341f3d53a3e175a3c92bd